+++ lib/src/widgets/future/common_builder.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+typedef AkoFutureBuilder<T> = Widget Function(T data);